LayoutTests/platform/gtk/accessibility/aria-help-expected.txt

button
This tests that aria-help attribute works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.description is 'AXDescription: click here'
PASS successfullyParsed is true

TEST COMPLETE
